+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed crash when used for React prior 16.9 because of another internals for dependencies (context)
+- Improved object diff for update events
 
 ## 0.3.0 (September 8, 2021)